tthresh

Function compress

pub fn compress<T: Element>(
    data: &[T],
    shape: &[usize],
    target: ErrorBound,
    verbose: bool,
    debug: bool,
) -> Result<Vec<u8>, Error>

Compress the data buffer using the target error bound.

Errors

Errors with

Error::InsufficientDimensionality if the data’s shape is not at least three-dimensional

Error::InvalidShape if the shape does not match the data length

Error::ExcessiveSize if the shape cannot be converted into u32s

Error::NegativeErrorBound if the target error bound is negative